Introduce a uniform (hopefully) thread-safe FrameBuffer
The buffer structures have been a mix of vectors and high performance
pre-allocated pointer lists.

The new class FrameBuffer tries to create a thread-safe buffer
that can be accessed by multiple producers (writers) and possibly also by
multiple consumers (readers).

The design is similar to the one that was already present in
UDPThread since FrameBuffer uses three lists (pool, buffer and
intermediateBuffer) to organise and frames.

Signed-off-by: Maximilian Güntner <maximilian.guentner@gmail.com>