Added CSV Support, Timeouts are now in microseconds
cannelloni now accepts a csv file (-T file.csv) in the
format

CAN_ID,Timeout(us)\n

Example:

1,2000
3,300000

would set the individual timeout for frames with the id 1 to 2000 us
and the individual timeout for frames with the id 3 to 300000 us

UDPThread:
If a timeout for an ID is lower than the current timer value,
the timer value will be changed to that timeout.
This can be used to send high priority frames.

Furthermore, all timers except the CANThread timer have been
changed to microseconds and helper functions have been
introduced.

Signed-off-by: Maximilian Güntner <maximilian.guentner@gmail.com>