Improve writing CAN frames in CANThread
until now, CANThread just tried to write all frames which often
resulted in packet loss.
This tries to deferr sending frames once the write() call blocks.
Instead of using FrameBuffer's swapBuffers(), we simply extract single
frames from the front() of the buffer and either put it back
into the pool if the write() was successful or we return it to the front
of the buffer, so the next call to transmitBuffer can try again.

Signed-off-by: Maximilian Güntner <maximilian.guentner@gmail.com>