Add CAN FD support
CAN FD frames are now being handled on interfaces that
support CAN FD.
If a CAN FD frame is sent to a cannelloni instance that
has only a CAN 2.0 interface, the frame will get dropped
and a warning is displayed since this most likely a
configuration error.

Thanks to @hartkopp for the hint.

Signed-off-by: Maximilian Güntner <maximilian.guentner@gmail.com>