Also send CAN FD flags
When a CAN FD frame has been, the MSB of frame->len
will be set and when transmitted using UDP, frame->flags
will be inserted between the length information and the
data section.
Prior to transmitting, the MSB in frame->len will be cleared.
CAN FD frames are dropped on 'legacy' sockets (CAN 2.0).

This raises CANNELLONI_FRAME_VERSION to 2.

Signed-off-by: Maximilian Güntner <maximilian.guentner@gmail.com>