Efficiency: Don't send partial frames
If the FrameBuffer was >= one network (UDP or SCTP) packet,
transmitFrame() did just send the remaining frames in
another UDP/SCTP packet/chunk.
This often resulted in one big packet (~ 1400 Byte) and one
small packet (~ 18 Byte).
Since the last frames in the buffer either are the newest ones
or the ones with the lowest priority (high ID), we can safely
put them back into the buffer and they are sent with the next
UDP packet.

This solution is a bit ugly but does the job - maybe there is
a better method to this, though.

Signed-off-by: Maximilian Güntner <maximilian.guentner@gmail.com>