Efficiency: Stop the timer when framebuffers are empty
When the framebuffers are empty, there is no need to
poll at a high rate, therefore the timer can be disabled
until a frame is inserted.

In order to shutdown UDPThread/SCTPThread even if the
transmit timer is disabled, a new timer has been introduced
that times out every 500ms.
The advantage of this method compared to using a select timeout
is that we can trigger/change this timer from the main program.

Signed-off-by: Maximilian Güntner <maximilian.guentner@gmail.com>